# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Meticulous API",
    "description": "REST API for the Meticulous CI/CD platform",
    "contact": {
      "name": "Meticulous Team"
    },
    "license": {
      "name": "UNLICENSED"
    },
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "/",
      "description": "Current server"
    }
  ],
  "paths": {
...
    "/api/v1/agents": {
      "get": {
        "tags": [
          "agents"
        ],
        "operationId": "list_agents",
        "parameters": [
          {
            "name": "status",
            "in": "query",
            "description": "Filter by agent status",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "pool",
            "in": "query",
            "description": "Filter by pool",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "tags",
            "in": "query",
            "description": "Filter by tags (comma-separated)",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "cursor",
            "in": "query",
            "description": "Pagination cursor",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Items per page",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of agents",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/v1/agents/{id}": {
      "get": {
        "tags": [
          "agents"
        ],
        "operationId": "get_agent",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Agent ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Agent details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Agent not found"
          }
        }
      },
      "delete": {
        "tags": [
          "agents"
        ],
        "operationId": "delete_agent",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Agent ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Agent removed from organization"
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Agent not found"
          },
          "409": {
            "description": "Agent still has running jobs"
          }
        }
      }
    },
    "/api/v1/agents/{id}/drain": {
      "post": {
        "tags": [
          "agents"
        ],
        "operationId": "drain_agent",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Agent ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Agent draining",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentActionResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Agent not found"
          }
        }
      }
    },
    "/api/v1/agents/{id}/resume": {
      "post": {
        "tags": [
          "agents"
        ],
        "operationId": "resume_agent",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Agent ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Agent resumed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentActionResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden"
          },
          "409": {
            "description": "Agent not in draining state"
          }
        }
      }
    },
    "/api/v1/agents/{id}/revoke": {
      "post": {
        "tags": [
          "agents"
        ],
        "operationId": "revoke_agent",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Agent ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Agent revoked",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AgentActionResponse"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden"
          },
          "404": {
            "description": "Agent not found"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AgentActionResponse": {
        "type": "object",
        "required": [
          "agent_id",
          "status",
          "message"
        ],
        "properties": {
          "agent_id": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "status": {
            "type": "string"
          }
        }
      },
      "AgentResponse": {
        "type": "object",
        "required": [
          "id",
          "org_id",
          "name",
          "status",
          "tags",
          "os",
          "arch",
          "version",
          "max_jobs",
          "running_jobs",
          "available_capacity",
          "created_at"
        ],
        "properties": {
          "arch": {
            "type": "string"
          },
          "available_capacity": {
            "type": "integer",
            "format": "int32"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "id": {
            "type": "string"
          },
          "last_heartbeat_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "last_security_bundle": {
            "type": "object",
            "description": "Snapshot of registration-time host / security fields (JSON object)."
          },
          "max_jobs": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string"
          },
          "org_id": {
            "type": "string"
          },
          "os": {
            "type": "string"
          },
          "pool": {
            "type": [
              "string",
              "null"
            ]
          },
          "pool_tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "running_jobs": {
            "type": "integer",
            "format": "int32"
          },
          "status": {
            "type": "string"
          },
          "tags": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "version": {
            "type": "string"
          }
        }
      },
...
    }
  },
...
}